+++ core/src/main/kotlin/net/corda/core/node/PluginServiceHub.kt
@@ -2,7 +2,6 @@ package net.corda.core.node
 
 import net.corda.core.flows.FlowLogic
 import net.corda.core.identity.Party
-import net.corda.core.identity.PartyAndCertificate
 
 /**
  * A service hub to be used by the [CordaPluginRegistry]
@@ -10,5 +9,5 @@ import net.corda.core.identity.PartyAndCertificate
 interface PluginServiceHub : ServiceHub {
     @Deprecated("This is no longer used. Instead annotate the flows produced by your factory with @InitiatedBy and have " +
             "them point to the initiating flow class.", level = DeprecationLevel.ERROR)
-    fun registerFlowInitiator(initiatingFlowClass: Class<out FlowLogic<*>>, serviceFlowFactory: (PartyAndCertificate) -> FlowLogic<*>) = Unit
+    fun registerFlowInitiator(initiatingFlowClass: Class<out FlowLogic<*>>, serviceFlowFactory: (Party) -> FlowLogic<*>) = Unit
 }