+++ docs/source/changelog.rst
@@ -43,9 +43,9 @@ UNRELEASED
     * There is a new ``AbstractParty`` superclass to ``Party``, which contains just the public key. This now replaces
       use of ``Party`` and ``PublicKey`` in state objects, and allows use of full or anonymised parties depending on
       use-case.
-    * A new ``PartyAndCertificate`` class has been added which contains an X.509 certificate and certificate path back
-      to a network trust root. This is widely used in place of ``Party`` inside Corda, with the exception of a few cases
-      where proof of identity is not required.
+    * A new ``PartyAndCertificate`` class has been added which aggregates a Party along with an X.509 certificate and
+      certificate path back to a network trust root. This is used where a Party and its proof of identity are required,
+      for example in identity registration.
     * Names of parties are now stored as a ``X500Name`` rather than a ``String``, to correctly enforce basic structure of the
       name. As a result all node legal names must now be structured as X.500 distinguished names.